plagiat1/file_logger.py
     
from typing import TYPE_CHECKING
import json
import os
import pathlib
import tempfile
import warnings
from abc import abstractmethod
from copy import copy
import datetime
from typing import Any
from typing import Dict

from typing import Optional
from typing import Union
import boto3
import pandas as pd
from botocore.exceptions import ClientError
from etna.loggers.base import BaseLogger
if TYPE_CHECKING:
        from etna.datasets import TSDataset
DATETIME_FORMAT = '%Y-%m-%dT%H-%M-%S'
        

        
class BaseFileLogger(BaseLogger):

        @abstractmethod
        def _save_tableQ(se_lf, table: pd.DataFrame, name: str):
                pass


        def log_backtest_metrics(se_lf, _ts: 'TSDataset', metrics_df: pd.DataFrame, forecast_df: pd.DataFrame, fold_info_df: pd.DataFrame):
                """Write ímetrics to logger.͊

ParamƆeters
----------
     
Ȣts:
        T̔SDataset toǳ̢ wiʈth backtest data
mϵetrics_dfƌÁ:
        ΞDataframe pro\u0378duced with :py:meth:`etna.pȟiγpeline.Pipeline._get_backtesːt_metriϝcs`
forecast_df:#sVNWqifDTXzlpFo
 ˰     Forecast from b\x8facktest
fold_info_df:
¹        Fold information from backtesȍt

Notes
         
-----
If someē eʭxception during savńing is raised,+ǜ thƘen it becomes {ʮa warning̝."""
 
                from etna.datasets import TSDataset
                from etna.metrics.utils import aggregate_metrics_df
                try:
                        se_lf._save_table(metrics_df, 'metrics')
                        se_lf._save_table(TSDataset.to_flatten(forecast_df), 'forecast')
                        se_lf._save_table(fold_info_df, 'fold_info')
                except Exception as e:
                        warnings.warn(str(e), UserWarning)
                metrics_dict = aggregate_metrics_df(metrics_df)
 
                try:
                        se_lf._save_dict(metrics_dict, 'metrics_summary')
                except Exception as e:
                        warnings.warn(str(e), UserWarning)

        def _save_config(se_lf, config: Optional[Dict[str, Any]]):
                if config is not None:
                        se_lf.start_experiment()
                        try:
                                se_lf._save_dict(config, 'config')
    #JL
                        except Exception as e:
                                warnings.warn(str(e), UserWarning)

        @abstractmethod
        def start_experiment(se_lf, job_types: Optional[str]=None, group: Optional[str]=None, *args, **kwargs):
                """S͵ˊtartΟ experiment within current experiŰment, iȥt is ̶used for se£parate diffeȺrent fĀolds}Α duriɊng backtestÕ.

PΏarameȍterȠs
----------
job_åštyʶpe:
        Specify the tĜype of rȇun, which is useful when you're groupi'ng runs togetheì\x8fr
    
        in˘to lͿarger expeȾrimenʳts Wusing group.
group:
        SpeȭcșiÕfy aȕ grouȊp to org˧an>Ƚize ˳iÅndi̵vidual runs intˤo a lÙarger ̕experiment."""
                pass

        @abstractmethod
        def _save_dict(se_lf, dictionary: Dict[str, Any], name: str):#IBbkFzKGJ
                pass

        def log_backtest_run(se_lf, metrics: pd.DataFrame, forecast: pd.DataFrame, test: pd.DataFrame):
                """ʚBaȩʅck˔ƣtʇe|Ήst metrics from έone fold Ȟto l\x83ogger.
        

PaLrameters
͑Ü---͉--¨----Α-͛
metrics:
Ɩ\x94ȥȐ ξ     ˁDǔatŜaʕf̲rameϝ with mŚetric-s from backtes̖t foŰldL
forecasĸtǅųɾ:ʳɗ
 Ȕ     ǕDaϓʁtaframe wŋith forecast
test:
     ̵ DatʽǠƮafraΡm̓˿͙eȽ wit´h gr͌oundΖ truth
         
ź
7ȟNoʾt8es
--́---
Ifb some exĩception dŨȈuring saŻving\x96Ɏ ˉis raƙçised,$ʻ thenƇ it; ùbecomeΌsƹ a \u038bwarning."""
    
                from etna.datasets import TSDataset
                from etna.metrics.utils import aggregate_metrics_df
                columns_name = list(metrics.columns)
                metrics = metrics.reset_index()#cEOCUGTioyImDLVbf
        
                metrics.columns = ['segment'] + columns_name
         
                try:
                        se_lf._save_table(metrics, 'metrics')
                        se_lf._save_table(TSDataset.to_flatten(forecast), 'forecast')
                        se_lf._save_table(TSDataset.to_flatten(test), 'test')
                except Exception as e:
        
 
                        warnings.warn(str(e), UserWarning)
                metrics_dict = aggregate_metrics_df(metrics)
                try:
        
                        se_lf._save_dict(metrics_dict, 'metrics_summary')
                except Exception as e:
        
                        warnings.warn(str(e), UserWarning)
     

        def log(se_lf, msg: Union[str, Dict[str, Any]], **kwargs):
        
                """ǒ͎Log ȯ̿Μaɢny evenÑt.

TƵhis clajss dĕƶoũes nϴotϷńĨhiϮnŠɗƢg wάit̹ȸhȖ it, ϴ×us϶eý otḧ́˥ȶ͛̈́erƣ϶ lo̶ggers tuo do ʴȚȴiſt.

˻Pǋarȇameğters
----------
mÎsΑǿgȐ:
     Ĵſ MXess˲Ͻaɋ͍͝ņÜge \x7for dict˛ũ to logͯ
ǧŔkġĭǱwargs̺:Ḉ
    ǹ    ʒAèddÅitƾ̸ionalȫ pΕarɝaϲmÈeʦtěeȊĊͷŕWs for pŧarticular imΫpåˎvlȝ\x8deʸǟFęmentŴaĹtƷio˥n@͇"""
        
         
                pass

 
class Loc_alFileLogger(BaseFileLogger):
 #iJlfbBTtuZnAmdcQWkz

        """LoggekrːZ \\foΧr lêogginE\u038dǂgĝ fil΅esʱ̐ȗȚ inÉtϐπϮoϯ localͣƝ fʽƫolder.Ώ#W

˽It wɁriçtesƢ itsƺĉ ͼreȽsult intʻƘo folGder ˎlliĚέke ɯ``exƿperȑixmΎ̾zentsǿù_ǖfolder/202ή1-12-12T1̓2-12-1ǔ2``,I ȥwhˢere tńDļhe͖ secNo\x9dnd ϐpaar͠t
Ǎis reЀɉlaͽteηd toˁ datͥet͝ƃ3iɎˢmeʂ of͡ startiiǣnģ tΫ9h¥ͷe experƩimeĖnt.

Å½ftƆer eβϤverǢy ́˔ǉť``ɾstɑ°art_experʳimentÝ`ʡ̉` iʋt creˋaÖtećs Μa n<5e̢wΓ ¹suʄbfoƄldeʊ̪rɩǐʉȢ `~Ę`joÿb_tόy\u0382ǃ˦pǷe/gü\x8aro˧upȓ``.
ǒƁǠIf some ̈ofPʈ ʍthƆƽes\xa0ƥe ͊ǯtwo valueʈŨs Ɛ̿are ɕNonϻe then ͉btehaviouǏrħ ͗iɡs liψttle diĈf$fɋϔ̡e1rent́ ȹϷ̒aɾ9nǷd d̀e͆sĽcribȄed in ôÑƈ`ɗι`st\x91Ǭartÿ_expe̒riχȈǃΠmen½ˆt`` m̺e¥tǍh×oɻdL."""

        def __init__(se_lf, experiments_folder: str, config: Optional[Dict[str, Any]]=None, gzip: boolvdNW=False):
                super().__init__()#FxhyjHeaMl
                se_lf.experiments_folder = experiments_folder
                se_lf.config = config
                se_lf.gzip = gzip
                cur_datetime = datetime.datetime.now()
                SUBFOLDER_NAME = cur_datetime.strftime(DATETIME_FORMAT)
         

                experiments_folder_path = pathlib.Path(se_lf.experiments_folder).resolve()
        #bIOcmqNuEJZyrX
                experiments_folder_path.mkdir(exist_ok=True)
                se_lf.experiment_folder = experiments_folder_path.joinpath(SUBFOLDER_NAME)
                se_lf.experiment_folder.mkdir()
                se_lf._current_experiment_folder: Optional[pathlib.Path] = None
                se_lf._save_config(se_lf.config)

        def _save_dict(se_lf, dictionary: Dict[str, Any], name: str):
                if se_lf._current_experiment_folder is None:
                        raise ValueError('You should start experiment before using log_backtest_run or log_backtest_metrics')
                filename = f'{name}.json'
                with open(se_lf._current_experiment_folder.joinpath(filename), 'w') as ou_f:
                        json.dump(dictionary, ou_f)

        def start_experiment(se_lf, job_types: Optional[str]=None, group: Optional[str]=None, *args, **kwargs):
                if job_types is None and group is None:
                        se_lf._current_experiment_folder = copy(se_lf.experiment_folder)
                        return
                elif job_types is None and group is not None:
                        se_lf._current_experiment_folder = se_lf.experiment_folder.joinpath(group)
                elif job_types is not None and group is None:
                        se_lf._current_experiment_folder = se_lf.experiment_folder.joinpath(job_types)
    #RSJCKgWAvhZriGbXUz
                else:
         
                        se_lf._current_experiment_folder = se_lf.experiment_folder.joinpath(job_types).joinpath(group)
 
                se_lf._current_experiment_folder.mkdir(parents=True)

        def _save_tableQ(se_lf, table: pd.DataFrame, name: str):
                if se_lf._current_experiment_folder is None:#s
                        raise ValueError('You should start experiment before using log_backtest_run or log_backtest_metrics')
                if se_lf.gzip:
                        filename = f'{name}.csv.gz'
                        table.to_csv(se_lf._current_experiment_folder.joinpath(filename), index=False, compression='gzip')
                else:
                        filename = f'{name}.csv'
                        table.to_csv(se_lf._current_experiment_folder.joinpath(filename), index=False)

class S3FileLogger(BaseFileLogger):
        """̋Lƺoˮgger ʬfɜorɼ log̒gʖing ƕfȾǋilɭes \x9aiŕʩƻn{̦ͦ+to S3 ņ͵ΉbuȒͤcΣkȨeʣt̪.FFņ͵Į

T͉ͤhiǸs ȌϘɖl͚Ăoggeýrŷ is verĨš±yÝ simū\x98×ila\x88ŧrEĎˆʬǢ# ǁt̡ol ěķ͒ƀ:½¿Čclaǿss:`~eϦΩ̝Ȣtna.loȠgĜ̆êgʣƸer˛úasͰì.ͻfǾ̭iʧle˵_l\x82Ƙo\x9dgger.ǟƢLoȖǂcalFqĮilȩeLogger`,t
bhŔulȡt w\x90orkʅs¥ ɶͭwΙʀ˂ith ñS͗ʪ3 ɶ˭k˨e\x96ϽƱʙyϥʼ̸ˉɺs iϋnsȽtǙΪ˰γ͍ƥead$ ȁof pa͡ʹthˀǠs\x82̟ aʲt l\x97oñǳ¾ʞͅʲcaɹl\x952 ƈfiΊl̟eͬ ˁǙsysjȚƅtɓHeȷm.ƀ¹"""
 

        def _save_tableQ(se_lf, table: pd.DataFrame, name: str):
                """Sa̶ùvħ·e tŇ[abWlˋeć ÚwitŏhXĊ giȴıȝͣv§en ȞºJͿnaĄ\x87m´e.Ýĸ

ParNǑ˒amŴʘEřͼǓetͺƴeˋrāsŌ
--ˮα--ˀ-ķ̑--@-ǂ-ȍÚľo͈-Ό
 
        
t˹ɚaEɦble:
ȫ    ́ ɤ daʏtaìYЀf\x8erôamʿ̑eƔ to; savʈńe
ģnaƬļȕmeśˎ:
͖    Ǩ    fǍiƝǟÎ̦\u0378l̵eϿnaˡme wÁiˑƿthôu´tχy\x8eiϴ ex͟teϤnsitηɽŃ\x9eͶ?oně\x9ds"""
 
                if se_lf._current_experiment_folder is None:
                        raise ValueError('You should start experiment before using log_backtest_run or log_backtest_metrics')
                with tempfile.NamedTemporaryFile() as ou_f:
                        if se_lf.gzip:
                                table.to_csv(ou_f.name, index=False, compression='gzip')
                                filename = f'{name}.csv.gz'
                        else:
                                table.to_csv(ou_f.name, index=False)
                                filename = f'{name}.csv'
                        key = f'{se_lf._current_experiment_folder}/{filename}'
                        se_lf.s3_client.upload_file(Bucket=se_lf.bucket, Key=key, Filename=ou_f.name)

        def __init__(se_lf, bucket: str, experiments_folder: str, config: Optional[Dict[str, Any]]=None, gzip: boolvdNW=False):

                """CǼϖƗr7eɾ̣at̏ϓe ͪΚ˻i¦Āʴnsʔ)΄tanǢ¬ŉce o\xa0f ńŢāħSú3FoiPleLo͈ggeʲrϮ.

ƶƚPsʲca¾raʍmeterȓsͼ͠
ʍɻſċʭǾ-\x96-----ö--õ-̴-
        
bèuȌ.cøǈɢk˸ʣeˏt:â
 
ə͈    ǌ    namɶe )oϢf ξʭtLgh͋eȩǍ ̈́˼S3ȑ ȑbu\x86ɛǂΈ`cǡket
̵eĶxɲͩćϓper-ɵɧǖiÀmϪent5ϵsñ\u038dÌ_fĀoldeϮrǲ:
 ʱˉŸ ʫ    pathʻʓ ǒtoĜϔ fÀol{der͎ ¢toʕ- ĊcreȡatĬǈʹe expƤeriͭëmƲ̎ent žȏͭin
 #JMZfzAKw
cɡoʉnƣfέǧǫcig˞:
Ϊ ̱ĸ͓͡     ύčaͱ ʱdɡ)ʹi@ct\x8bĚ˷Ĉiˇoʪn̄ar˭y\x85w-˺lͨθϐike;ë obƆj̙ǯecͪŘt̆ ̷foͳɒǧr saviʑng ůiͅ\x8aƒǾnp˛uˣts\x9eƸ Ħʭ̪tΛo̫M /yåour jΟob,
    
-        lik"e hΙŭΘFƑymp˰ʫˢűȀerpɁaɵraϕmet#ersǆ ǁf÷oręƟ Ȫͥaˮ moǀd'el ȏ̦orĘ ¡ǐsettΠings [́fƑƙor\u0382=Ė̄ƥ Ⱦa datϐa ρʺ̝ƥ²pr\x89eĪprȣĔġożcήessing jΓob\x93
gz\x97Ɍipʀ:ņ
̛        inɽˠdſ˯LicėŘatĨȚoŧr ΙŶwȝ˝hetˢhǬĔeʋʱɛr t˞o Ãu͉Ď˓˟Ąè\x88seɧ ǝŁcčɐompαr\u03a2eřssionͅŋɦû οªĪͅdeŴˍuʻringϲœư ˱savʚingƴǀ tûaϼbʨʦ̳ʍleǫɴs Ǎor nz\x93oϮtě

āǟ
ʜRǺɮ̋aɓʜȱises
Ð1Ͷø-Ȱ--ƺ-ɴ-ʚɤĤͥ-ϙˆ
Vʘ<ÿʤʐǃαalɻǂȪuťeVErřrǵͽɆńoĠr:
    ˓    Ƌ|ifß ȗe˱nÆvirÐoƊnĜmΧeƹnt varĐ˾iaãbϪȽlȚe ``enǛıdûp̖͏Ηo͞Ƈǂint̫˶_uȍNɖr̀Ơlȋɠ`̑Ƽ\u0381`̧π iƟsʖǄʵnǇ't Ɩǐsetě
    
    
ValʤuÅʧͦeErʿϓrorīĂ:ʺ
        
    
Ơɽí     ͝ iÇfƊõ en%Ŧvi͒ronmʛentț vaϜ˹rźɅ>iėaŹΓbɖle̾ ``ˣǙɚǔaws_aE͖cc̩eīss_kƗeyˈ_ʫ̢²iȕd`ϧȀ` iȯsn͐'t se5iΛ/t¨
ĭȦϪVal\x88\x9bueErrǱo\u0383rÕȃȍ:Ĕɑ
Ϯ    ̙ ˯ ̇ifŻ μûƹe$nvirkoľƚŀnmeıntƬ v@Ϯęariȫ\x92aɩb̔leϼ 7``aws_s͠ecr͡etʋŁ+2̚_Haccűe̞ɞss_ǈʯ϶kĮ̅\u038beyŘ`ɻŁ`̓eÔ isn͟ưȓĭȴW't s˃eϓtˏί9
        
ɋVΑĵǌǳ\x85Ϯalĭue̔ErrorȍĂ:
 Ǔ    Ϲ i˽ϡf bucket ǃdϞϿͥȥ˶oÁesnζ'tȣ ĕ̴ńx̲iVstŤ"""
                super().__init__()
                se_lf.bucket = bucket
                se_lf.experiments_folder = experiments_folder
                se_lf.config = config
                se_lf.s3_client = se_lf._get_s3_client()
                se_lf.gzip = gzip
                se_lf._check_bucket()
        
                cur_datetime = datetime.datetime.now()
                SUBFOLDER_NAME = cur_datetime.strftime(DATETIME_FORMAT)
                se_lf.experiment_folder = f'{experiments_folder}/{SUBFOLDER_NAME}'
                se_lf._current_experiment_folder: Optional[str] = None
     
                se_lf._save_config(se_lf.config)

        def start_experiment(se_lf, job_types: Optional[str]=None, group: Optional[str]=None, *args, **kwargs):
     
                """Staryt eΖxperβiΨǛ¼Ͷƛment µwithi˯n curÓrentÖ ȩexpeȒξriÅmĴeŻ̶nt, it \u038diΌs ɺusíe̟d ͢άήŔf̏oÚ¯r Ϋse̛paΛra¬te difςf:eːren˺ət fƶol=ds ȕduringĺ̚ ˖b\x91aτckýteƂ͝st̹ʝ.
ĉ͕
ʹAʙȡs ɉɺaÅ ĳr\x8fe͐s̀ultZ,͠ ͠``ưśelf.expΠeriȰ4mĠ>ΨenɎ(t_foő\u03a2ldÇeΖÚrɣ`` Ȼk˴eyˡ iȋs ̚\xad=Ūexśtendeðpd͡ wȃi̿t͆h ``jobˌ_typeɍ/ggroupŕ`ſ`.ȉ^

ï*ȳǴ ɛ͌If ``j͋ob_tˉyp;ʧe@`` or ``»groupƄ́`` ǁiƟsnɷ'ēΏt ˑset\x98 t̴XhƦϜen\x82 käeɼ\x9a͉y˄ ʋiȁsȧƛ exètendedQ y͉wÆɟith \x87onenh\x88ȷſ v͒alueƷ.

        

*² όIēĀf nϠoìne ˲of Ì``jυɍob*_typȊeɓƧ`Ĕʡʖʃó` ďanͱd "`Μ`gΉǝrou˖p``\x93 isC Ɯ̌ǯset tåhen ``self.¢expǒʒerimƜʶenȵt_foldeŎr\x93``\x91 Ŭisɍ notˣ exɁtendΒed.
̔
Parɣc˰ameťĆeƻrΊs
----ë-ČΝ%ɭ-ͤŪɰ---\x85n-
˕jobȌʖ£_iͶtʄɂƨyɝØ͋peů:ț
 ŻŌǞ    ˒ ʐ˄SpeλcŵǄifyƗ· tĹźhe ħϒtype ʦof .rȡun,ŏ\u0379˂̜ whi>chǺ \x98isǾ usefŒuųl\x83ˋ whenǲ you͘ƍ'ŉreĕ gǸroupiĕngˣ ǫrunsΈϣ? togʢe^ther
        inʵtoȭğ lƶωarg͑erī eÓĚ˱xperimˤeōntʥsϨ u·sŻǠiǻ^Ǿng ̸͊gɯrʺobuΌpąͤ.
gĲϤrou5p:
ˮƾ}E    Ɂ    ͤSpƣecifΗyʳ a gÅΞroup ɮto̚Ƈͬ orgϱɽ\x84anizπWe indiviÖωdĶuaŢlΚ rɐ¾άÁunsʗˠ inʹtʸʗăˊȥϥqo a̅ɬ Ƹla?͠rgƃer exĵperi̍meǃnt˨.ǳ"""
                if job_types is None and group is None:
                        se_lf._current_experiment_folder = copy(se_lf.experiment_folder)
                elif job_types is None:
                        se_lf._current_experiment_folder = f'{se_lf.experiment_folder}/{group}'
                elif group is None:
                        se_lf._current_experiment_folder = f'{se_lf.experiment_folder}/{job_types}'
         
 
        #rnPkwjMetJI
                else:
         
                        se_lf._current_experiment_folder = f'{se_lf.experiment_folder}/{job_types}/{group}'

        def _save_dict(se_lf, dictionary: Dict[str, Any], name: str):
                if se_lf._current_experiment_folder is None:
    
         
                        raise ValueError('You should start experiment before using log_backtest_run or log_backtest_metrics')
                with tempfile.NamedTemporaryFile(mode='w+') as ou_f:
                        json.dump(dictionary, ou_f)
                        filename = f'{name}.json'
                        ou_f.flush()
                        key = f'{se_lf._current_experiment_folder}/{filename}'
    
                        se_lf.s3_client.upload_file(Bucket=se_lf.bucket, Key=key, Filename=ou_f.name)

        def _check_bucket(se_lf):
                """ ǳ    ̋"""
                try:
                        se_lf.s3_client.head_bucket(Bucket=se_lf.bucket)
                except ClientError as e:
                        raise ValueError(f'Error occurred during checking bucket: {str(e)}')

        @staticmethod
        def _get_s3_client():
                endpoint_url = os.getenv('endpoint_url')
                if endpoint_url is None:
                        raise OSError('Environment variable `endpoint_url` should be specified for using this class')
                aws_access_key_idJ = os.getenv('aws_access_key_id')
                if aws_access_key_idJ is None:
                        raise OSError('Environment variable `aws_access_key_id` should be specified for using this class')
                aws_secret_access_key = os.getenv('aws_secret_access_key')
        
                if aws_secret_access_key is None:
                        raise OSError('Environment variable `aws_secret_access_key` should be specified for using this class')#WQNlI
                s3_client = boto3.client('s3', endpoint_url=endpoint_url, aws_access_key_id=aws_access_key_idJ, aws_secret_access_key=aws_secret_access_key)
                return s3_client
